+++ Controllers/CustomersController.cs
@@ -149,71 +149,5 @@ namespace PositronAPI.Controllers
         {
             return await _loyaltyService.GetLoyaltyCard(customerId);
         }
-
-        // Coupon Endpoints
-
-        /// <summary>
-        /// Add a coupon for customer
-        /// </summary>
-        /// <remarks>Creates a coupon.</remarks>
-        /// <param name="body">Properties for creating a new coupon.</param>
-        [HttpPost]
-        [Route("/customer/{customerId}/coupon")]
-        public async Task<ActionResult<Coupon>> CreateCoupon([FromBody][Required] Coupon body)
-        {
-            var response = await _couponService.CreateCoupon(body);
-
-            if (response == null) { return BadRequest(); }
-
-            return Ok(response);
-        }
-
-        /// <summary>
-        /// Remove a coupon
-        /// </summary>
-        /// <remarks>Deletes a coupon.</remarks>
-        /// <param name="couponId">The id of the coupon</param>
-        [HttpDelete]
-        [Route("/customer/{customerId}/coupon/{couponId}")]
-        public async Task<ActionResult> DeleteCoupon([FromRoute][Required] long couponId)
-        {
-            var response = await _couponService.DeleteCoupon(couponId);
-
-            if (response == null) { return NotFound(); }
-
-            return NoContent();
-        }
-
-        /// <summary>
-        /// Get a coupon of a customer
-        /// </summary>
-        /// <remarks>Gets a coupon.</remarks>
-        /// <param name="couponId">The id of the coupon</param>
-        [HttpGet]
-        [Route("/customer/{customerId}/coupon/{couponId}")]
-        public async Task<ActionResult<Coupon>> GetCoupon([FromRoute][Required] long couponId)
-        {
-            var response = await _couponService.GetCoupon(couponId);
-
-            if (response == null) { return NotFound(); }
-
-            return Ok(response);
-        }
-
-        /// <summary>
-        /// Gets coupons of a customer
-        /// </summary>
-        /// <remarks>Gets coupons.</remarks>
-        /// <param name="customerId">The id of the coupon holder</param>
-        [HttpGet]
-        [Route("/customer/{customerId}/coupon")]
-        public async Task<ActionResult<List<Coupon>>> GetCoupons([FromRoute][Required] long customerId)
-        {   
-            var response = await _couponService.GetCoupons(customerId);
-
-            if (response.Count == 0) { return NoContent(); }
-
-            return Ok(response);
-        }
     }
 }
